fix : test condition
