+++ src/main/java/com/syu/cara/user/controller/KakaoOAuthController.java
@@ -7,6 +7,8 @@ import com.syu.cara.user.dto.KakaoUserInfoDTO;
 import com.syu.cara.user.repository.UserRepository;
 import com.syu.cara.user.security.JwtService;
 import com.syu.cara.user.service.KakaoClient;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -72,5 +74,18 @@ public class KakaoOAuthController {
                 .build();
 
         return ResponseEntity.ok(responseDto);
+    }
+
+    @PostMapping("/logout")
+    public ResponseEntity<Void> logout(HttpServletResponse response) {
+        // JWT를 HttpOnly 쿠키에 저장하는 경우에만 쿠키를 삭제합니다.
+        Cookie cookie = new Cookie("jwtToken", null);
+        cookie.setHttpOnly(true);
+        cookie.setPath("/");
+        cookie.setMaxAge(0);
+        response.addCookie(cookie);
+
+        // 클라이언트 로컬 스토리지나 세션 스토리지는 클라이언트 측에서 삭제
+        return ResponseEntity.noContent().build();
     }
 }
